web-animations/animation-player/animation-player-source.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationPlayer source attribute return the source content associated with this player or null</title>
<meta name="assert" content="AnimationPlayer source attribute return the source content associated with this player or null">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-source">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var player = document.timeline.play();

    assert_equals(player.source, null, 'AnimationPlayer.source attribute value ' +
        'should be null');
}, 'Test AnimationPlayer.source attribute value is null if player ' +
    'is created by call to AnimationTimeline.play() with no arguments');

test(function() {
    var player = document.timeline.play(null);

    assert_equals(player.source, null, 'AnimationPlayer.source attribute value ' +
        'should be null');
}, 'Test AnimationPlayer.source attribute value is null if player ' +
    'is created by AnimationTimeline.play(null) call');

test(function() {
    var player = document.timeline.play(new Animation(createDiv(this), [], 5));

    assert_not_equals(player.source, null, 'AnimationPlayer.source attribute value ' +
        'should not be null');
}, 'Test AnimationPlayer.source attribute value is not null if player ' +
    'is created by AnimationTimeline.play(animation) call');

test(function() {
    var player = document.timeline.play(new Animation(createDiv(this), [], 5));

    assert_true(hasAncestorClassString(player.source, 'AnimationNode'),
        'AnimationPlayer.source attribute value should be of type AnimationNode');
}, 'Test AnimationPlayer.source attribute value is of type AnimationNode');

test(function() {
    var animation = new Animation(createDiv(this), [], 5);
    var player = document.timeline.play(animation);

    assert_equals(player.source, animation, 'AnimationPlayer.source ' +
        'attribute value should be the value passed to the AnimationTimeline.play()');
}, 'Test AnimationPlayer.source attribute value is the value passed ' +
    'to the AnimationTimeline.play()');
</script>
</body>
